Split coven blob I/O out of manager/sync.rs
manager/sync.rs mixed two layers: thin delegation to SyncController and
coven blob I/O (blob-refs, byte reads, pin state, cache budgets). Move the
coven blob surface verbatim into a new manager/coven_blobs.rs so sync.rs is
only the SyncController delegation veneer.
